+++ webpack.config.js
@@ -1,9 +1,9 @@
 var Encore = require('@symfony/webpack-encore');
 
 Encore
-    .setOutputPath('src/Resources/public/js/')
+    .setOutputPath('src/Resources/public/assets/')
     .addEntry('contao-multilingual-fields-bundle', './src/Resources/assets/js/contao-multilingual-fields-bundle.js')
-    .setPublicPath('/public/js/')
+    .setPublicPath('/public/assets/')
     .disableSingleRuntimeChunk()
     .enableSassLoader()
     .configureBabel(function (babelConfig) {